requirements-test.txt
Django==1.6.3
PyYAML==3.11
astroid==1.4.2
colorama==0.3.5
django-activity-stream==0.6.0
django-filter==0.11.0
django-kronos==0.8
djangorestframework==3.1.1
dodgy==0.1.9
lazy-object-proxy==1.2.1
logilab-common==1.1.0
mccabe==0.3.1
pep257==0.7.0
pep8==1.6.2
pep8-naming==0.3.3
pluggy==0.3.1
prospector==0.11.7
py==1.4.31
pyflakes==1.0.0
pylint==1.5.2
pylint-celery==0.3
pylint-common==0.2.2
pylint-django==0.7.1
pylint-flask==0.1
pylint-plugin-utils==0.2.3
requirements-detector==0.4.1
setoptconf==0.2.0
six==1.10.0
tox==2.3.1
virtualenv==13.1.2
wrapt==1.10.6
wsgiref==0.1.2
